LayoutTests/imported/w3c/web-platform-tests/fetch/images/canvas-remote-read-remote-image-redirect-expected.txt

CONSOLE MESSAGE: line 18: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.

PASS Load a no-cors image from a same-origin URL that redirects to a cross-origin URL that redirects to the initial origin 
